Separated core, ui, ktor, okhttp modules.
