+++ gradle.properties
@@ -1,6 +1,7 @@
 #Kotlin
 kotlin.code.style=official
 kotlin.daemon.jvmargs=-Xmx3072M
+kotlin.mpp.applyDefaultHierarchyTemplate=false
 
 #Gradle
 org.gradle.jvmargs=-Xmx4096M -Dfile.encoding=UTF-8
