+++ lens-core/src/desktopMain/kotlin/io/github/farhazulmullick/lenslogger/Platform.desktop.kt
