--- lens-logger/src/androidMain/AndroidManifest.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-    <application>
-        <provider
-            android:name="io.github.farhazulmullick.lenslogger.plugin.network.LensContextProvider"
-            android:authorities="${applicationId}.lenslogger.contextprovider"
-            android:exported="false" />
-    </application>
-</manifest>
