+++ lens-logger/src/commonMain/kotlin/io/github/farhazulmullick/lenslogger/LensLogger.kt
@@ -1,0 +1,8 @@
+package io.github.farhazulmullick.lenslogger
+
+/**
+ * Marker for the composite `lens-logger` artifact. Prefer explicit dependencies on
+ * `lens-core`, `lens-ktor`, `lens-ui`, and `lens-okhttp` when you only need part of the stack.
+ */
+@Suppress("unused")
+object LensLogger
