+++ lens-ui/src/commonMain/composeResources/drawable/logger_no_data.xml
