+++ lens-ui/src/commonMain/composeResources/drawable/no_data_prefs.xml
